refactor: rename JobModel.ts to jobQueryModel.ts
