+++ server/src/models/JobQueryModel.ts
