+++ server/src/server.ts
@@ -7,7 +7,7 @@ import gptAPIRouter from './routes/api/gptRoutes.js';
 import favoriteRoutes from './routes/favorites.js';
 import jSearchRoutes from './routes/api/jSearchRoutes.js';
 import sequelize from './config/connection.js';
-import JobModel from './models/JobModel.js';
+import JobModel from './models/JobQueryModel.js';
 
 
 dotenv.config();
